add blocks different buildTime

+++ src/game/blocks/block_map.hpp
@@ -118,7 +118,7 @@ public:
     //
     void place(TileCoord tile, TeamID teamID, std::unique_ptr<Block>& block);
     void demolish(TileCoord tile);
-    void build(const TileCoord tile, const TeamID teamID, const int8_t buildSpeed, const Presets& presets, Inventory& inventory);
+    void build(const TileCoord tile, const TeamID teamID, const int16_t buildSpeed, const Presets& presets, Inventory& inventory);
     void applyBlueprint(const Blueprint& blueprint, const TeamID teamID, const Presets& presets);
     std::optional<InProgressAim> getClosestInProgress(const PixelCoord mobPosition);
     //
